♻️ refactor: Create `generalUtils` for common utility functions
- Created `generalUtils.js` to centralize shared utility functions
- Moved existing utility function from `date.js` to `generalUtils.js`
