+++ js/utils/generalUtils.js
